[build] Add user.basic environment, use it for vDSO
This adds the generalized user.basic environment for building
things to use only the basic machine ABI. This is now used for
the vDSO, replacing the old special user.vdso environment. It
will later be used for things that need to avoid using the
Fuchsia Compiler ABI.

This restores the long-lost feature of doing variant selection
for the vDSO. It now gets independent variant selection using the
select_variant selector clauses matching the loadable_module()
with output_name="libzircon", so it doesn't necessarily match the
variant used for the kernel or for the rest of userland, but it
will for simple and compatible variants like "gcc".

Bug: 69697, 122877, 121817
Change-Id: I361a4e596098207f4b28c80d019c81363135484c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825536
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Alex Brachet <abrachet@google.com>