[dfv2] Fix loader under ASAN
In DFv2 the loader is manually installed in the DriverHost. This means
it does not go through the normal process setup, and so DriverManager
must manually configure the loader for ASAN.

Also, the Compat driver should not be cloning the loader, as a clone of
the loader loses configuration information.  Instead, update the Compat
driver to use an unowned connection to the existing loader.

Now that the loader was updated to make blocking calls, the backing
loader in the tests are run on a separate thread.

Fixed: 105603
Change-Id: Ifc1b6089d86bf580ccace192b8770895bfae7f7f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832781
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>