[starnix] Fix SelectTest.SetrlimitCallNOFILE
This test failure had nothing to do with RLIMIT. Instead, we were
failing this test because select for no events on a tmpfs file caused
and error instead of blocking until the timeout. Previously, select was
implemented in terms of epoll, which produces an EPERM error in this
situation.

Bug: 76815
Change-Id: Ic1c8293531ec2f8cfee0ef30095649c68f77e872
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832736
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>