[amlogic-display] Explain primary node choice and "colocate" requirement.
DsiImpl banjo protocol is not supported by fragment-proxy for
inter-devhost forwarding. So the request for DsiImpl client
of DsiHost in amlogic-display driver may not success when dsi-dw
driver (implementing DsiImpl protocol) and amlogic-display driver
are not running on the same devhost.

As a result, the "dsi" node of amlogic-display must be the
primary node and amlogic-display must be colocated with its
parent. However, `colocate` property in its component manifest
is only an advisory so this may break one day.

This adds some comments and TODOs explaining the reason of
primary node choice and `colocate` requirement, and notes that
this is very brittle and can be subject to failure.

Bug: 125227
Change-Id: I125fbb7cf2d1ee86a298f62f6304e7895a6732e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832749
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>