[syslog] Don't assume va_list is safe to use uninitialized
It's not really kosher to use va_list without va_start
intializing it. But in practice it's sufficient to explicitly
default-initialize it. It's never OK to just use an uninitialized
value, even just for passing it to be ignored elsewhere.

Bug: 122875
Change-Id: Icbe8332e09c9d6007373d1cc36ef8846c999113e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833318
Reviewed-by: Ankur Mittal <anmittal@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Paul Kirth <paulkirth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>