Reland "[vim3][power] Fix pwm and core drivers"
This is a reland of commit a509994cee6236c4ab38de4cadd5a6f659b4325f

In the original commit I had used the AddCompositeImplicitPbusFragment
function which seems to cause the colocation flag to mess up. This was
making it so that the power driver was not able to connect to the banjo
protocol it needed. Fixed this by manually putting the pdev fragment and
using the AddComposite function instead.

Tests: Tested with the device enumeration test on a vim3 dfv1 build
(originally I had only tested dfv2)

Original change's description:
> [vim3][power] Fix pwm and core drivers
>
> Fixes the aml_pwm_regulator and its child power driver.
>
> Bug: 124320
> Tests: device-enumeration test on vim3
>
> Change-Id: I6ccc9784385d759c6173e145a25c14c270dde83a
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831493
> Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
> Commit-Queue: Novin Changizi <novinc@google.com>

Bug: 124320
Change-Id: Ia3ea9a735b94e714dfccade1c6c5c4bebe43e75c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832836
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>