[vfs][rust] Replace directory internals with io2
The beginning of implementing Open2. The general idea is to replace io1
types such as fuchsia.io/OpenFlags with generics that are convertible to
io2 types, and then implement the logic in terms of io2.

This commit cheats by bringing the gap above and below the affected
directory APIs using DirectoryOptions::into_io1, but it is intended to
demonstrate how this API surface can be migrated gradually.

Remove `directory::common::test::new_connection_validate_flags_create`
because it's not clear what it means to create a new connection with
CREATE.

Adjust the existing implementations of GetConnectionInfo:
- OpenFlags.RIGHT_READABLE   maps to R_STAR_DIR
- OpenFlags.RIGHT_WRITABLE   maps to W_STAR_DIR
- OpenFlags.RIGHT_EXECUTABLE maps to X_STAR_DIR

Bug: 77623
Change-Id: Idfa9a22b9f0a1d1a30d32dd50b7114768a806a79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832877
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>