[boringssl] s/BSDish/OpenSSL for license tag
BSDish is a reasonable stab at the problem given the... free-wheeling...
nature of the LICENSE file. However, BSDish isn't actually a license,
and getting a more structured level of license tagging is a priority.

rhilary@ 's at OSPO approved doc for what licenses are and are not
concurs that OpenSSL is a reasonable tag for this unreasonable license
text. See
https://fuchsia.dev/fuchsia-src/contribute/governance/policy/open-source-licensing-policies?hl=en#approved-licenses

Change-Id: I072d4ae65cf737fda2d6fdf5b939133dddc616db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/834897
Commit-Queue: Chris Lewis <cflewis@google.com>
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>