[driver_manager] Don't reload Driver for firmware
The FirmwareLoader is able to get the driver struct out of the device
instead of re-loading it through the driver loader.

This simplifies the firmware loader class as it no longer needs to use
the coordinator. This means that the coordinator doesn't need to forward
declare the class.

TEST: ddk-firmware-test
Bug: 125288
Change-Id: I040934358c92e04ce64e4a05fb734ebd6c40de1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833856
Reviewed-by: Sarah Chan <spqchan@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>