[bedrock] Implement Remote for multishot::Sender
This makes it possible for Senders to be used over FIDL in programs,
and inserted into CloneDicts.

For example, a program can use this to connect to a FIDL protocol by
sending a server ends through the Sender to the serving Receiver.

Bug: 122024

Change-Id: I46f0d0aad2583dce347b37e2433f12346429d001
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833524
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
API-Review: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>