[amlogic-display] Move capture support check out of Vout.
As described in amlogic-display/README.md, amlogic display engine
can forward the VPP (blended OSD) output to the VOUT instance for
encoding and display, or forward it to the loopback interface for
readback to the memory.

Thus, whether a display engine supports capture is unrelated to the
type of Vout, so we should move it out of Vout.

Currently there's some issue on VIM3 where display capture tests
are not working; so we disabled capture support on VIM3 display
driver.

Bug: 124947
Test: On VIM3 with [case 1] TS050 touchscreen [case 2] HDMI display, run
      $ fx shell display-test --capture
      capture checks are disabled.
Change-Id: I77b33d5f60fdda99c98b70a9e6537f0650099999
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/834696
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>