[starnix] Fix flock_test
We now pass this whole test suite. The flock operation is better
modelled as an operation on the FileObject than the node itself. To pass
the last test case, we need to fail flock operations on O_PATH
FileObjects with EBADF. This factoring avoids the need to pass the
FileHandle to the FsNode, which is a bad code smell.

Fixed: 79272
Change-Id: I5e1159597f2bba28bb063927963424af9b5ba4da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835073
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>