[bedrock] Make multishot::Receiver cloneable
Switches mutishot to use async_channel Sender/Receiver instead
of futures::mpsc. `async_channel::Receiver` implements Clone.

This makes it possible to put Receivers in CloneDicts, which can be
used to implement a restartable FIDL protocol server. Upon restart,
the new instance of the server program receives a copy of the
CloneDict and receives protocol server ends through the Receiver.

Values are (still) not cloned, and will be delivered to only one copy
of the Receiver.

Bug: 122024

Change-Id: I5afbdce4547177ca1ff1edd180d987a2a7860d04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/834457
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>