[driver runner] Refactor bind code
Split the match driver callback into separate functions.
Due to increased scope and complexity, the logic in the
match callback became difficult to maintain and error
prone.

This CL moves the logic to bind the node into
`BindNodeToResult()` and handling the results in
`OnMatchDriverCallback()`.

This reduces the chance of forgetting to invoke a callback
for bind completion.

Bug: 124988
Tests: driver_runner_tests
Change-Id: I09a648edac3396a0490ee52e28b68f54c71a8d36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/834140
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>