[assembly] `create-update` can optionally rewrite default repo
This changes `ffx assembly create-update` to add a
`--rewrite-default-repo` argument that edit's the update package's
`packages.json` to use a different repository hostname. This was then
used to change scrutiny to use this to switch all the update package
urls to `test.fuchsia.com`.

This begins a graceful transition for the security pkg test to migrate
to passing in the `update_domain` rather than passing in a test
configuration.

Change-Id: I6799e2805d5e8a9f33600b81abfb145dce171ef8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833516
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>