[kernel] Add`and` command to kernel shell, use in unittest script
This adds a new meta-command to the kernel shell: `and ...` does
the ... command only if the last command succeeded.

This is used in the kernel-unittests-boot-test script along with
the `-r` feature of `ut` for the suites that previously the
userland invoker ran repeatedly.  This gets the kernel shell
script version of the unit tests back up to par with the previous
way to run them via the standalone core-tests userland boot test.

This also factors out the GN logic into a couple of templates to
make it easier to define boot tests based on kernel shell scripts.

Bug: 122154
Change-Id: I3601863463cc7fb538004f5d000eec852456bbc1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835359
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>