[bedrock] Add RestartableElf
RestartableElf is a proof of concept of a component-like object that
manages the execution of multiple Elf program instances over time.

It can be started and stopped, which recreates an Elf program from
a declaration (path to the executable) and an incoming Dict, which
contains capabilities provided to each program instance.

Tests demonstrate a FIDL client that reconnects when restarted,
and a restartable FIDL server that handles multiple connections.

Bug: 122024

Change-Id: I024dda99b4e1fe9dfcddb78ab63bd91d1e170851
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825375
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>