[kernel][riscv][mmu] Implement EnlargeOperation for Unmap ops
In the case of an unmap operation that fails to allocate memory while
splitting a large page table, the EnlargeOperation option signals
whether or not it's okay for the underlying code to fail by simply
dropping all of the mappings that large page covered. Now is parity with
the other arch mmu implementations.

Bug: 122875
Change-Id: I0280d0aa94df61677fa6155770808222454b8570
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835063
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>