[mock-mmio-reg] Hides MockMmioReg behind interface
Currently, the library requires users to create an array of MockMmioReg
instances only for the purposes of passing the array to the
MockMmioRegRegion's constructor (as a ptr). The tests themselves don't
really ever use the array itself. Also, the MockMmioRegRegion exposes
the underlying *Reg instances by way of operator[].

One complication is that this library is used across //integration repo
edges, which cannot be updated atomically. This adds an intermittent
backwards-compatible version. Once all the uses outside fuchsia.git are
updated, the compatibility shim will be removed (and MockMmioReg hidden
in a private namespace).

Bug: 122816
Change-Id: If9f0372c991d4867cfb939f56a8e751cdb92d2c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818728
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Randall Bosetti <rlb@google.com>
Commit-Queue: Sam Hansen <hansens@google.com>