[fake-mmio-reg] Hides FakeMmioReg behind interface
Currently, the library requires users to create an array of FakeMmioReg
instances only for the purposes of passing the array to the
FakeMmioRegRegion's constructor (as a ptr). The tests themselves don't
really ever use the array itself. Also, the FakeMmioRegRegion exposes
the underlying *Reg instances by way of operator[].

Bug: 122816
Change-Id: Ie8d3760024be627a05cc7bd8bc0aed03ae04f72e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821507
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Sam Hansen <hansens@google.com>