[driver][logging] Add node names to compat log tag
Adds the node names of all driver instances of a specific driver to the
log tags in the compat shim. This is useful because the component info
that CF shows may be incorrect in multi instance situations. This will
help the reader know all possible nodes that a log could be coming from.

Tests: Booted qemu and saw both single and multi instance drivers
logging successfully with the expected tags.

Change-Id: Ic0e4fe2bea68e9233e0af301d3a7e2851a194bdf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832700
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>