[kernel][riscv][mmu] loosen up the valid base/size checks for user aspaces
Some of the test code builds user aspaces that have different base and
lengths than are defined to be a user address space. This is fine from
a technical point of view of the mmu implementation, so loosen up the
code that checks for invalid sizes on aspace construction.

Bug: 122875
Change-Id: I4b3d203a57bbda9cb4c7500a7a0c989fd47a83bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/836384
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>