[fidl][fuchsia.boot.Arguments] Increase Collect result vector max
fuchsia.boot/Arguments.Collect was defined with a maximum return
value, regardless of how many boot arguments actually matched the
prefix, of 32 (MAX_ARGS_VECTOR_LENGTH) elements. This maximum,
however, is ignored by the server in component_manager, which returns
as many elements as there are matches. This behavior is used by
"ffx target show" which calls Collect with the empty string to obtain
all of the boot arguments, which exceeds 32 on some configurations
(the configuration that discovered this issue had 35).

This behavior was broken recently when length validation was added to
the Rust FIDL bindings. Now, instead of returning more than 32 boot
arguments, component_manager closes the connection.

This commit increases MAX_ARGS_VECTOR_LENGTH from 32 to 255
elements (which keeps the maximum return message size under the the 64
KiB limit and should be enough for anybody).

This is in a sense a breaking change for Collect, as clients built
before the change that were validating the return length will now
error when talking to a server built after the change that is sending
more than the old maximum. In practice, the only server of this API,
component_manager, was already sending more than the maximum and this
change undoes the effective API breakage introduced when the Rust
bindings started validating the length.

The other methods should not be affected because the server only
returns as many elements as the client requests.

Additionally, this FIDL API is internal so ffx is the only client we
need to worry about breaking.

Also, the length limit on the GetStrings return vector is changed from
MAX_ARGS_COLLECT_LENGTH (presumably a typo) to MAX_ARGS_VECTOR_LENGTH
so that clients can request up to the new maximum without error.

Bug: 125464
Change-Id: I17ae3615f04c372a96392315b5f350c60585a30e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/836716
Commit-Queue: Ben Keller <galbanum@google.com>
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
API-Review: Carlos Pizano Uribe <cpu@google.com>