[kernel][riscv] skip a lock dep regression test on single cpu runs
Skip instead of failing the bug #84827 regression test if booting with a
single core system. Though it's not ideal the test cannot be run it's
also not ideal to fail all of the unit tests because of the single core
situation.

Bug: 122875
Change-Id: I70216f36a191d62f6acd9acc16aee22de0aa76ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/836825
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>