[async-utils] Support keeping hanging-get server hydrated
...with an in-flight request.

This change updates `async_utils::hanging_get::client::HangingGetStream`
to be configured to eagerly send a (hanging-get) request so that the
server is able to serve a response as quickly as possible so the
`Future` is able to do meaningful work as soon as possible.

Before this change, a request was sent during a poll following:
  1) the creation of the `HangingGetStream`
  2) the completion of a previous in-flight request/`Future`

Now, a request is sent immediately after the above situations (before
polling) if configured to do so.

Change-Id: I3580232ffe6e56f9065d634d2900ef0d03e0b84c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837056
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>