[netcfg] Hold DHCPv4 client-ends in a `StreamMap`
This change introduces the first-steps in implementing the client-end
of the `WatchConfiguration` hanging-get. NetCfg how holds a stream of
DHCPv4 configurations in a `StreamMap` and eagerly sends a
`WatchConfiguration` request to the DHCPv4 client component.

Note that this change does not yet handle a response to the
`WatchConfiguration` request; that will come in a follow-up.

Bug: 125247
Change-Id: I6fdbf09015a448e5020ac312a37ca23f722727f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837077
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>