[magma] Split logging initialization into PlatformLoggingProvider
Initializing logging requires a dependency on PlatformHandle, which has
a dependency on several other libraries. It's also not necessary at all
with the DDK logger, so it makes sense to split it out into a separate
library that's only depended on when necessary.

Bug: 111471
Change-Id: I6358dd585b35c59d402ac6650106091d6f05dc36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833339
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: John Rosasco <rosasco@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>