[media][codec] Create base class for software codec.
Refactor so that common StreamProcessor / CodecImpl logic can be reused
by most software encoders and decoders.
ChunkInputStream is more commonly used for encoders. The change makes
use of the ChunkInputStream for processing CodecPackets easily, but
makes it so that the time extrapolator functionality is only used for
encoders.

Test: fx test //src/media/stream_processors/Test

Change-Id: I0b84303c247af1b7eeb3b3a5694a222ab01ddd87
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/806107
Commit-Queue: Dayeong Lee <dayeonglee@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>