[fxfs] Fixed a flush/attributes race
There was a small window where it was possible to get old attributes
whilst a flush was taking place.

The fix is to check the in-memory attributes before retrieving the
attributes from the device (rather than the other way round which is
what opens up the window).

Fixed: 125544
Change-Id: I52770c22c0fe2c571e2cdbbfa0863e13442e3890
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838716
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Rui Qi Sim <rqsim@google.com>