[botanist] Write blobs into tempfile in same directory as destination
Previously we would write files into the system temp directory, which
might not be on the same file system as the botanist blob cache. This
changes botanist to create a temp directory in the same directory as
the destination blob to make renaming the blob to be atomic.

This has the risk that this could leave behind tempfiles in the cache
directory if the go process is SIGKILL-ed, but should be cleaned up
with SIGINT.

Change-Id: Ic641447edf9410a089040b2eca5abf820824715d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838498
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ina Huh <ihuh@google.com>